requirements.txt
# --- Core ---
PyYAML>=6.0         # per caricare/parlare con i file .yaml
requests>=2.31.0    # utile per chiamate HTTP (webhook notifiche)
tqdm>=4.66.0        # barre di progresso opzionali
psutil>=5.9.0       # monitoraggio risorse CPU/memoria locali

# --- Logging & Monitoring ---
prometheus_client>=0.17.0  # per integrazione base con Prometheus exporter

# --- Report generation ---
pandas>=2.2.0        # per elaborare metriche
matplotlib>=3.9.0    # per grafici/report PDF
reportlab>=4.0.9     # per generare report PDF automatici

# Scientific stack (tested compatible)
numpy>=1.26.4,<2.0
scipy>=1.11.0
pandas>=2.2.0
matplotlib>=3.8.0

# --- Optional HPC interaction ---
# slurm CLI commands (sbatch, squeue, scancel) are already available in MeluXina,
# so NO Python library needed, subprocess will handle them

# --- Development ---
pytest>=8.0.0        # per test
black>=24.0.0        # per formattazione codice
